+++ src/app.d.ts
@@ -1,6 +1,8 @@
 // See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
 declare global {
+  const __NPM_PACKAGE_VERSION__: string;
+
   namespace App {
     // interface Error {}
     // interface Locals {}