+++ vite.config.ts
@@ -49,5 +49,8 @@ export default defineConfig({
         }
       }
     ]
+  },
+  define: {
+    __NPM_PACKAGE_VERSION__: JSON.stringify(process.env.npm_package_version || "")
   }
 });